+++ CHANGELOG.md
@@ -14,12 +14,22 @@ any parts of the framework not mentioned in the documentation should generally b
 
 * Add support for Django 2.2
 
+### Changed
+
+* Allow to define `select_related` per include using [select_for_includes](https://django-rest-framework-json-api.readthedocs.io/en/stable/usage.html#performance-improvements)
+* Reduce number of queries to calculate includes by using `select_related` when possible
+
 ### Fixed
 
 * Avoid exception when trying to include skipped relationship
 * Don't swallow `filter[]` params when there are several
 * Fix DeprecationWarning regarding collections.abc import in Python 3.7
-* Allow OPTIONS request to be used on RelationshipView.
+* Allow OPTIONS request to be used on RelationshipView
+
+### Deprecated
+
+* Deprecate `PrefetchForIncludesHelperMixin` use `PreloadIncludesMixin` instead
+* Deprecate `AutoPrefetchMixin` use `AutoPreloadMixin` instead
 
 ## [2.7.0] - 2019-01-14